obtain lock before broadcaster.publish to avoid race
when calling `with broadcaster` we invoke it's __aenter__ and __aexit__
before calling publish(). publish() assumes a trasaction to the underlying
engine was established and is active. (i.e: asyncpg.connect()).

when two parallel `with broadcaster` and `publish` are racing each
other, one __aexit__ will close the transaction, causing the publish()
query to fail.