Broadcaster's disconnect test works
Removed postgres service in favor of running the postgres db container from within the test.
Changed test_ws_closes_on_pg_broadcaster_disconnect to validate messages are still broadcasted after idle timeout in DB connection.
(PubSubClient is responsible for retrying the connection, so waiting on it to finish was wrong)